+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0rc2 (unreleased)
 ---------------------
 
+- #1643 Setup View Filter
 - #1642 Allow multi-choice in results entry
 - #1640 Fix AttributeError on Worksheet Template assignment
 - #1638 Fix "Published results" tab is not displayed to Client contacts